05-smart-robot/requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
baidu-aip==4.16.13
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
dashscope==1.23.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
edge-tts==7.0.2
erniebot==0.5.9
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.0
future==1.0.0
gradio==5.23.1
gradio-client==1.8.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.9.1
mcp-server==0.1.4
mcp-server-sqlite==2025.4.25
mdurl==0.1.2
mistune==3.1.3
modelscope-studio==1.1.7
mpmath==1.3.0
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook-shim==0.2.4
numpy==1.26.4
onnx==1.18.0
onnxruntime-cann==1.23.0
openai==1.82.0
opencv-python==4.11.0.86
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patchelf==0.17.2.2
pdfminer-six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
playsound==1.3.0
prometheus-client==0.22.0
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyaudio==0.2.14
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.9.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.3
pypdfium2==4.30.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
qwen-agent==0.0.25
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.11
safehttpx==0.1.6
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
send2trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
speechrecognition==3.14.3
srt==3.5.3
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.143
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
uv==0.7.6
uvenv==3.9.3
uvicorn==0.34.2
uvx==2.5.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
xlsxwriter==3.2.3
yarl==1.20.0